fix bashrc

+++ bashrc
@@ -1,9 +1,14 @@
-source /usr/share/defaults/etc/profile
-# source /usr/lib/python2.7/site-packages/powerline/bindings/bash/powerline.sh
+#
+# ~/.bashrc
+#
+
+# If not running interactively, don't do anything
+[[ $- != *i* ]] && return
 
 # duso aliases
 alias su='su -l'
 alias sudo='sudo '
+alias sudoe='sudo -e'
 alias cp='cp -v'
 alias mv='mv -vi'
 alias rm='rm -v'
@@ -11,16 +16,12 @@ alias mkdir='mkdir -p -v'
 alias ..='cd ..'
 alias ...='cd ../..'
 
-alias sudoe='sudo -e'
-alias eopkg='eopkg -v'
-alias reboot='sudo reboot'
-alias poweroff='sudo poweroff'
 
 alias ping='ping -c 5'
 alias mkdir='mkdir -p'
-alias grep='grep --color=auto'
 
+alias grep='grep --color=auto'
 alias ls='ls --color=auto'
 
-# run stuff
-# neofetch
+# shell prompt
+PS1='[\u@\h \W]\$ '
